% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceLLama.R
\name{ceLLama}
\alias{ceLLama}
\title{Annotate Clusters}
\usage{
ceLLama(
  marker.list,
  n_genes = 20,
  seed = 101,
  base_prompt =
    "Act like an expert immunologist and give me the cell type annotation for this cluster. Please, reply only with the answer and nothing else! If you're not sure just label it as 'unsure'.",
  model = "llama3",
  get_reason = FALSE,
  url = "http://localhost:11434/api/generate",
  temperature = 0.1
)
}
\arguments{
\item{marker.list}{A list of data frames containing gene markers for each cluster.}

\item{n_genes}{Number of top genes to select for both up- and down-regulated lists.}

\item{base_prompt}{A base prompt to use for formatting the annotation data.}

\item{model}{The model to use for annotation.}

\item{get_reason}{boolean. Re-prompts LLM for the annotation reason.}

\item{url}{The URL for the annotation service.}

\item{temperature}{a number between 0-1, creates diversity for the response}
}
\value{
A list of annotations and reasons for each cell cluster.
}
\description{
This function annotates cell clusters based on their top up-regulated and down-regulated genes.
}
