% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_html_report}
\alias{create_html_report}
\title{Function to convert Markdown to HTML}
\usage{
create_html_report(md_file = "report.md", html_file = "report.html")
}
\description{
Function to convert Markdown to HTML
}
