% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceLLama.R
\name{format_annotation_data}
\alias{format_annotation_data}
\title{Format Annotation Data}
\usage{
format_annotation_data(top_genes, base_prompt)
}
\arguments{
\item{top_genes}{A list of top genes for each cluster.}

\item{base_prompt}{A base prompt to use for formatting the annotation data.}
}
\value{
A list of formatted annotation prompts for each cluster.
}
\description{
Format Annotation Data
}
