% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report_md}
\alias{generate_report_md}
\title{Function to generate a simple report in Markdown format}
\usage{
generate_report_md(annotations, output_file = "report.md")
}
\arguments{
\item{annotations}{results of ceLLama, cell type annotations}
}
\description{
Function to generate a simple report in Markdown format
}
