% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceLLama.R
\name{get_annotation}
\alias{get_annotation}
\title{Get Annotation for a Cluster}
\usage{
get_annotation(description, model, url, seed = 100, temperature = 0)
}
\arguments{
\item{description}{The description for the cluster to be annotated.}

\item{model}{The model to use for annotation.}

\item{url}{The URL for the annotation service.}
}
\value{
The annotation for the cluster.
}
\description{
Get Annotation for a Cluster
}
