% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceLLama.R
\name{get_reason}
\alias{get_reason}
\title{Get Reason for the Annotation}
\usage{
get_reason(
  annotation,
  description,
  model,
  url,
  seed = 100,
  temperature = 0,
  verbose = T
)
}
\arguments{
\item{annotation}{The annotation received for the cluster.}

\item{description}{The description used to generate the annotation.}

\item{model}{The model to use for the reason generation.}

\item{url}{The URL for the reason generation service.}
}
\value{
The reason for the annotation.
}
\description{
Get Reason for the Annotation
}
