% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceLLama.R
\name{select_top_genes}
\alias{select_top_genes}
\title{Select Top Up- and Down-Regulated Genes}
\usage{
select_top_genes(degs, n = 20)
}
\arguments{
\item{degs}{A data frame containing gene markers with their log2 fold changes.}

\item{n}{Number of top genes to select for both up- and down-regulated lists.}
}
\value{
A list with top up-regulated and down-regulated genes.
}
\description{
Select Top Up- and Down-Regulated Genes
}
